application/requirements.txt
aiohttp==3.8.4
aioredis==2.0.1
aiosignal==1.3.1
alembic==1.11.1
amqp==5.1.1
annotated-types==0.5.0
anyio==3.6.2
argilla==1.12.0
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
backoff==2.2.1
bcrypt==4.0.1
billiard==4.1.0
celery==5.3.1
certifi==2023.5.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
clickhouse-connect==0.6.6
coloredlogs==15.0.1
commonmark==0.9.1
cryptography==38.0.4
dataclasses-json==0.5.8
Deprecated==1.2.14
dnspython==2.2.1
docx2txt==0.8
duckdb==0.8.1
ecdsa==0.18.0
email-validator==1.3.0
et-xmlfile==1.1.0
exceptiongroup==1.1.2
faiss-cpu==1.7.4
fastapi==0.88.0
filetype==1.2.0
flatbuffers==23.5.26
frozenlist==1.3.3
greenlet==2.0.2
h11==0.14.0
hnswlib==0.7.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.1
humanfriendly==10.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.1
jq==1.4.1
kombu==5.3.1
langchain==0.0.207
langchainplus-sdk==0.0.16
loguru==0.7.0
lxml==4.9.3
lz4==4.3.2
Mako==1.2.4
Markdown==3.4.3
MarkupSafe==2.1.1
marshmallow==3.19.0
marshmallow-enum==1.5.1
monotonic==1.6
mpmath==1.3.0
msg-parser==1.2.0
multidict==6.0.4
mypy-extensions==1.0.0
nltk==3.8.1
numexpr==2.8.4
numpy==1.23.5
olefile==0.46
onnxruntime==1.15.1
openai==0.27.8
openapi-schema-pydantic==1.2.4
openpyxl==3.1.2
orjson==3.8.3
outcome==1.2.0
overrides==7.3.1
packaging==23.1
pandas==1.5.3
passlib==1.7.4
pdf2image==1.16.3
pdfminer.six==20221105
pgvector==0.1.8
Pillow==10.0.0
pinecone-client==2.2.2
posthog==3.0.1
prompt-toolkit==3.0.39
protobuf==4.23.4
psycopg2-binary==2.9.6
pulsar-client==3.2.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.9
pydantic_core==2.1.2
Pygments==2.15.1
pypandoc==1.11
pypdf==3.10.0
PySocks==1.7.1
python-dateutil==2.8.2
python-decouple==3.8
python-docx==0.8.11
python-dotenv==0.21.0
python-jose==3.3.0
python-magic==0.4.27
python-multipart==0.0.5
python-pptx==0.6.21
python-slugify==8.0.1
pytube==15.0.0
pytz==2023.3
PyYAML==6.0
regex==2023.6.3
requests==2.31.0
rfc3986==1.5.0
rich==13.0.1
rsa==4.9
selenium==4.10.0
sentry-sdk==1.28.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.16
SQLAlchemy-Utils==0.41.1
sqlparse==0.4.4
starlette==0.22.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.2
text-unidecode==1.3
tiktoken==0.4.0
tokenizers==0.13.3
tqdm==4.65.0
trio==0.22.1
trio-websocket==0.10.3
typer==0.7.0
typing-inspect==0.9.0
typing_extensions==4.6.3
tzdata==2023.3
ujson==5.6.0
unstructured==0.8.0
urllib3==2.0.3
uvicorn==0.20.0
uvloop==0.17.0
vine==5.0.0
watchfiles==0.18.1
wcwidth==0.2.6
websockets==10.4
wrapt==1.14.1
wsproto==1.2.0
xlrd==2.0.1
XlsxWriter==3.1.2
yarl==1.9.2
youtube-transcript-api==0.6.1
zipp==3.16.0
zstandard==0.21.0
